Averages Usually Arent Ints
This commit changes the average efficiency field from an int to double because why would that ever be an int

+++ Destiny.ScrimTracker.Logic/Models/Guardian.cs
@@ -4,6 +4,6 @@ namespace Destiny.ScrimTracker.Logic.Models
     {
         public string GamerTag { get; set; }
         public int Elo { get; set; }
-        public int AverageEfficiency { get; set; }
+        public double AverageEfficiency { get; set; }
     }
 }
